工程项目源代码/requirements.txt
# Voice Recognition and Synthesis
pyaudio==0.2.11
baidu-aip==4.16.10
websocket-client==1.6.4
pygame>=2.5.2

# GUI Framework - Qt5
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0

# Web Service
Flask==3.0.0
Flask-CORS==4.0.0

# Scheduled Tasks
schedule==1.2.0
apscheduler>=3.10.4

# Audio Processing
playsound==1.3.0

# AI API (DeepSeek)
openai>=1.12.0

# Other Tools
requests==2.31.0
python-dotenv>=0.19.0

# Chinese Text Processing
# jieba>=0.42.1  # 已移除，使用优化的LLM语义识别